make it robust (#98)
* update controller
* frontend bug fix

+++ src/backend/src/main/java/se/sjtu/formfoss/controller/FormDataController.java
@@ -8,13 +8,13 @@ import org.springframework.web.bind.annotation.*;
 import se.sjtu.formfoss.exception.ObjectNotFoundException;
 import se.sjtu.formfoss.exception.PermissionDenyException;
 import se.sjtu.formfoss.model.FormDataEntity;
+import se.sjtu.formfoss.model.FormEntity;
 import se.sjtu.formfoss.repository.FormDataRepository;
-import se.sjtu.formfoss.exception.GlobalException;
+import se.sjtu.formfoss.repository.FormRepository;
+import se.sjtu.formfoss.util.AuthRequestUtil;
+import se.sjtu.formfoss.util.RestResponseUtil;
+
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 /**
  * Created by QHZ on 2017/7/4.
@@ -23,7 +23,10 @@ import java.util.Map;
 @RequestMapping(path = "${url.authentication}")
 public class FormDataController {
     @Autowired
-    FormDataRepository formDataRepository;
+    private FormDataRepository formDataRepository;
+
+    @Autowired
+    private FormRepository formRepository;
 
     //Get all form data
     @GetMapping(path="/formdata")
@@ -36,51 +39,42 @@ public class FormDataController {
 
     //Get form data by formId
     @GetMapping(path="/formdata/{fid}")
-    public @ResponseBody
-    ResponseEntity<FormDataEntity> getFormDataById(@PathVariable int fid,
-                                                   @RequestAttribute Integer userId,
-                                                   @RequestAttribute String userRole){
+    @ResponseStatus(code = HttpStatus.OK)
+    public @ResponseBody FormDataEntity getFormDataById(@PathVariable int fid,
+                                                        @RequestAttribute Integer userId,
+                                                        @RequestAttribute String userRole){
         FormDataEntity formData = formDataRepository.findOne(fid);
         if(formData==null){
             throw new ObjectNotFoundException("form data not found");
         }
 
-        FormDataEntity result=new FormDataEntity();
-        result.setFormId(formData.getFormId());
-        result.setAnswerCount(formData.getAnswerCount());
-        result.setData(formData.getData());
-        List<Map<String,Object>> data=formData.getData();
-        List<Map<String,Object>> subData=new ArrayList<Map<String, Object>>(formData.getData());
-        for(int i=0;i<data.size();i++){
-            Map<String,Object> answer=data.get(i);
-            if(answer.get("type").equals("textbox")){
-                List<String> textResult= (List<String>) answer.get("result");
-                if(textResult.size()>15){
-                    Map<String,Object> subAnswer=new HashMap<String, Object>(answer);
-                    answer.put("hasRestAnswer",true);
-                    data.set(i,answer);
-                    List<String> subTextResult=textResult.subList(0,15);
-                    subAnswer.put("result",subTextResult);
-                    subData.set(i,subAnswer);
-                }
-            }
+        FormEntity form = formRepository.findOne(fid);
+        // check privilege
+        if (!AuthRequestUtil.checkFormOwnership(form, userId, userRole)) {
+            throw new PermissionDenyException("can not access");
         }
-        formData.setData(data);
-        formDataRepository.save(formData);
-        result.setData(subData);
-        return new ResponseEntity<>(result,HttpStatus.OK);
+
+        return formData;
     }
 
 
 
     //Delete form data by formId
     @DeleteMapping(path="/formdata/{fid}")
-    public @ResponseBody ResponseEntity<String> delFormDataById(@PathVariable int fid){
+    @ResponseStatus(code = HttpStatus.NO_CONTENT)
+    public @ResponseBody String delFormDataById(@PathVariable int fid,
+                                                @RequestAttribute Integer userId,
+                                                @RequestAttribute String userRole) {
+        FormDataEntity data = formDataRepository.findOne(fid);
+        if (!AuthRequestUtil.checkFormDataOwnership(data, userId, userRole)) {
+            throw new PermissionDenyException("No privilege to access the data");
+        }
         formDataRepository.delete(fid);
-        return new ResponseEntity<String>("{\"message\": \"Delete successfully\"}",HttpStatus.OK);
+        return RestResponseUtil.successMsg("deleted");
     }
 
-    //Create form data
+    // Create form data
+    // deprecated
     @PostMapping(path="/formdata")
     public @ResponseBody ResponseEntity<String> createFormData(@RequestBody FormDataEntity formData) throws IOException {
         int fid=formData.getFormId();
@@ -98,6 +92,7 @@ public class FormDataController {
 
 
     //Update from data
+    // deprecated
     @PutMapping(path="/formdata")
     public @ResponseBody ResponseEntity<String> updateFormData(@RequestBody FormDataEntity formData){
         formDataRepository.save(formData);
